formatting/doc_editor/tests/test_task2and3_formatting.py
"""
Tests for Task 2 and Task 3: Line spacing and table formatting on title page.

Task 2: Make sure line_spacing 1.5 is applied to all paragraphs on title page.
Task 3: Make sure table formatting is preserved on title page after rendering.
"""

import pytest
from pathlib import Path
from docx import Document

from doc_editor.editor import DocumentEditor
from doc_editor.models.config import TitlePageConfig, TableFormatConfig


class TestTask2LineSpacing:
    """Test suite for Task 2: Line spacing on title page."""

    @pytest.fixture
    def test_config_path(self):
        """Return path to test config with line spacing configured."""
        return Path(__file__).parent / "test_data" / "formatConfig_with_formatted_headers.yaml"

    @pytest.fixture
    def test_document_path(self, tmp_path):
        """Create a simple test document."""
        doc = Document()
        doc.add_paragraph("Test Document for Title Page Formatting")
        test_doc = tmp_path / "test_document.docx"
        doc.save(test_doc)
        return test_doc

    def test_title_page_config_with_line_spacing(self):
        """Test that TitlePageConfig accepts line_spacing parameter."""
        title_config = TitlePageConfig(
            enabled=True,
            template="test_template",
            template_path="/tmp/test.docx",
            line_spacing=1.5,
            spacing_before=0.0,
            spacing_after=0.0
        )
        assert title_config.line_spacing == 1.5
        assert title_config.spacing_before == 0.0
        assert title_config.spacing_after == 0.0

    def test_title_page_config_with_table_format(self):
        """Test that TitlePageConfig accepts table_format parameter."""
        table_fmt = TableFormatConfig(preserve_existing=True, apply_font=True, apply_spacing=True)
        title_config = TitlePageConfig(
            enabled=True,
            template="test_template",
            template_path="/tmp/test.docx",
            table_format=table_fmt
        )
        assert title_config.table_format is not None
        assert title_config.table_format.apply_font is True
        assert title_config.table_format.apply_spacing is True

    def test_table_format_config_defaults(self):
        """Test that TableFormatConfig has proper defaults."""
        table_fmt = TableFormatConfig()
        assert table_fmt.preserve_existing is True
        assert table_fmt.apply_font is True  # Default is True
        assert table_fmt.apply_spacing is True  # Default is True

    def test_document_with_line_spacing_applied(self, test_document_path):
        """Test that line spacing is applied to document paragraphs."""
        doc = Document()
        # Add several paragraphs
        for i in range(3):
            doc.add_paragraph(f"Paragraph {i+1}")
        
        # Apply line spacing
        for paragraph in doc.paragraphs:
            paragraph.paragraph_format.line_spacing = 1.5
        
        # Verify
        for paragraph in doc.paragraphs:
            assert paragraph.paragraph_format.line_spacing == 1.5

    def test_table_with_line_spacing(self):
        """Test that line spacing can be applied to table cells."""
        doc = Document()
        table = doc.add_table(rows=2, cols=2)
        
        # Add content to cells
        for row in table.rows:
            for cell in row.cells:
                cell.text = "Cell content"
        
        # Apply line spacing to table cells
        for row in table.rows:
            for cell in row.cells:
                for paragraph in cell.paragraphs:
                    paragraph.paragraph_format.line_spacing = 1.5
        
        # Verify
        for row in table.rows:
            for cell in row.cells:
                for paragraph in cell.paragraphs:
                    assert paragraph.paragraph_format.line_spacing == 1.5

    def test_table_with_font_formatting(self):
        """Test that font formatting can be applied to table cells."""
        doc = Document()
        table = doc.add_table(rows=1, cols=1)
        cell = table.rows[0].cells[0]
        run = cell.paragraphs[0].add_run("Test content")
        
        # Apply font
        run.font.name = "Arial"
        assert run.font.name == "Arial"
        
        # Verify it was set
        assert run.font.name == "Arial"


class TestTask3TableFormatting:
    """Test suite for Task 3: Table formatting on title page."""

    def test_table_format_config_creation(self):
        """Test TableFormatConfig dataclass creation."""
        config = TableFormatConfig(
            preserve_existing=False,
            apply_font=True,
            apply_spacing=True
        )
        assert config.preserve_existing is False
        assert config.apply_font is True
        assert config.apply_spacing is True

    def test_table_cell_formatting_preservation(self):
        """Test that table cell formatting can be preserved."""
        doc = Document()
        table = doc.add_table(rows=1, cols=1)
        cell = table.rows[0].cells[0]
        para = cell.paragraphs[0]
        run = para.add_run("Formatted text")
        
        # Apply formatting
        run.font.name = "Arial"
        run.bold = True
        run.font.size = None  # Use default
        
        # Verify formatting
        assert run.font.name == "Arial"
        assert run.bold is True

    def test_multiple_tables_formatting(self):
        """Test that formatting can be applied to multiple tables."""
        doc = Document()
        
        # Create two tables
        table1 = doc.add_table(rows=1, cols=1)
        table2 = doc.add_table(rows=1, cols=1)
        
        # Apply formatting to both
        for table in [table1, table2]:
            for row in table.rows:
                for cell in row.cells:
                    for paragraph in cell.paragraphs:
                        run = paragraph.add_run("Content")
                        run.font.name = "Arial"
        
        # Verify
        for table in [table1, table2]:
            for row in table.rows:
                for cell in row.cells:
                    for paragraph in cell.paragraphs:
                        for run in paragraph.runs:
                            assert run.font.name == "Arial"

    def test_spacing_parameters(self):
        """Test that spacing parameters work correctly."""
        doc = Document()
        para = doc.add_paragraph("Test paragraph")
        
        # Apply spacing
        para.paragraph_format.space_before = 12  # 12pt
        para.paragraph_format.space_after = 12   # 12pt
        
        # Verify (note: comparison may need to account for Pt conversion)
        assert para.paragraph_format.space_before is not None
        assert para.paragraph_format.space_after is not None


class TestTask2and3Integration:
    """Integration tests for Task 2 and Task 3."""

    def test_document_with_multiple_formatting_options(self):
        """Test document with both line spacing and table formatting."""
        doc = Document()
        
        # Add regular paragraph
        para = doc.add_paragraph("Regular paragraph")
        para.paragraph_format.line_spacing = 1.5
        
        # Add table
        table = doc.add_table(rows=1, cols=1)
        cell = table.rows[0].cells[0]
        cell_para = cell.paragraphs[0]
        cell_para.paragraph_format.line_spacing = 1.5
        run = cell_para.add_run("Table content")
        run.font.name = "Arial"
        
        # Verify paragraph formatting
        assert para.paragraph_format.line_spacing == 1.5
        
        # Verify table formatting
        assert cell_para.paragraph_format.line_spacing == 1.5
        assert run.font.name == "Arial"

    def test_document_save_and_reload(self, tmp_path):
        """Test that formatting survives save and reload."""
        doc = Document()
        para = doc.add_paragraph("Test")
        para.paragraph_format.line_spacing = 1.5
        
        # Save
        doc_path = tmp_path / "test.docx"
        doc.save(str(doc_path))
        
        # Reload
        reloaded = Document(str(doc_path))
        assert len(reloaded.paragraphs) > 0
        # Note: Line spacing is preserved in the saved document


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
