formatting/requirements.txt
docxcompose==1.4.0
docxtpl==0.20.0
Flask==3.1.1
python_docx==1.1.2
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.4
